# doc/api.yaml
openapi: 3.0.0
info:
  title: WasaText
  description: |
    This API allows users to interact with each other 
    on the WasaText app.
    Users can create groups, private chats, send text or image messages, 
    reply, forward or react them.
  version: 1.0.0
servers:
  - url: http://localhost:3000/v1
    description: Production server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: UUID
      description: |
        Use the user identifier returned by `/login` in the Authorization header.
        Example: `Authorization: Bearer abcdef012345`
  responses:
    UnauthorizedError:
      description: Missing or invalid Authorization header
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: "Unauthorized"
  schemas:
...
    ImageUpload:
      type: object
      properties:
        file:
          type: string
          format: binary
          description: >
            Image file to upload. Allowed types: jpeg, png, webp.
            Max size: 2MB.
      required:
        - file
security:
  - BearerAuth: []
paths:
...
  /me/photo:
    post:
      tags:
        - Users
      summary: Upload a profile image
      description:
        Upload a user's profile image. The image must be of jpeg, png or webp format and its size
        must be at most 2Mb.
      operationId: setMyPhoto
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/ImageUpload"
            encoding:
              file:
                contentType: image/jpeg, image/png, image/webp
      responses:
        "200":
          description: Profile photo uploaded successfully
          content:
            application/json:
              schema:
                description: A descriptive message of success and the new image url
                type: object
                properties:
                  message:
                    type: string
                    example: "Profile photo updated"
                    description: The success message
                  photoPath:
                    type: string
                    example: profiles/1.png
                    description: The newly profile photo path
        "400":
          description: Invalid input image
          content:
            application/json:
              schema:
                description: A descriptive message of different failures
                type: "object"
                properties:
                  message:
                    type: "string"
                    description: >
                      error:
                      * `First Error`  - No file uploaded.
                      * `Second Error` - Wrong file format. Only jpeg, png and webp are allowed.
                      * `Third Error`  - File too big. The image can't be bigger than 2Mb.
                    enum:
                      - no_file_uploaded
                      - wrong_file_format
                      - file_too_big
                    default: "Wrong file format. Only jpeg, png and webp are allowed."
        "401":
          $ref: "#/components/responses/UnauthorizedError"
...
  /groups/{id}/members:
    parameters:
      - name: id
        in: path
        required: true
        description: The group id
        schema:
          type: integer
    post:
      tags:
        - Groups
      summary: Add a new member to a group
      description: Any user belonging to a group can add other users to the group if they are not already there.
      operationId: addToGroup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              description: The username of the member to add
              type: object
              required:
                - userId
              properties:
                userId:
                  type: integer
                  description: user id
      responses:
        "200":
          description: Member added to the group successfully.
          content:
            application/json:
              schema:
                type: object
                description: A descriptive message of success
                example:
                  message: Member added to the group successfully

        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                description: A descriptive message of failure
                type: object
                properties:
                  message:
                    type: string
                    description: Forbidden
              example:
                message: Forbidden
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                description: A descriptive message of different failures
                type: object
                properties:
                  message:
                    type: string
                    description: >
                      error:
                      * `First Error` - User not found 
                      * `Second Error` - Group not found
                    enum:
                      - user_not_found
                      - group_not_found
        "401":
          $ref: "#/components/responses/UnauthorizedError"
    delete:
      tags:
        - Groups
      summary: Leave a group
      description: A user can remove themselves from a group they belong to, or if he is admin, remove another user. If the user who deletes himself is an admin, another user will be promoted to admi will be promoted to admin.
      operationId: leaveGroup
      responses:
        "204":
          description: User successfully removed from the group
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                description: A descriptive message of failure
                type: object
                properties:
                  message:
                    type: string
                    description: User not part of the group
              example:
                message: User not part of the group
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                description: A descriptive message of failure
                type: object
                properties:
                  message:
                    type: string
                    description: Group not found
              example:
                message: Group not found
        "401":
          $ref: "#/components/responses/UnauthorizedError"
  /groups/{id}/name:
    put:
      tags:
        - Groups
      summary: Change group name
      description: Update the group name
      operationId: setGroupName
      parameters:
        - name: id
          in: path
          required: true
          description: The group id
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              description: Update the group name
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  description: the new group name
            example:
              name: I ragazzi del calcetto
      responses:
        "200":
          description: Group name changed successfully.
          content:
            application/json:
              schema:
                description: A descriptive message of success
                type: object
                example:
                  message: Group name changed successfully.

        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                description: A descriptive message of failure
                type: object
                properties:
                  message:
                    type: string
                    description: >
                      error:
                      * `First Error` - Invalid group name 
                      * `Second Error` - Forbidden
                    enum:
                      - invalid_group_name
                      - forbidden
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                description: A descriptive message of failure
                type: object
                properties:
                  message:
                    type: string
                    description: Group not found
              example:
                message: Group not found
        "401":
          $ref: "#/components/responses/UnauthorizedError"
  /groups/{id}/photo:
    put:
      tags:
        - Groups
      summary: Change group photo
      description: Update the group photo
      operationId: setGroupPhoto
      parameters:
        - name: id
          in: path
          required: true
          description: The group id
          schema:
            type: integer
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/ImageUpload"
            encoding:
              file:
                contentType: image/jpeg, image/png, image/webp
      responses:
        "200":
          description: Group photo uploaded successfully
          content:
            application/json:
              schema:
                description: A descriptive message of success
                type: object
                properties:
                  message:
                    type: string
                    example: Group photo updated successfully
                    description: A success message
                  photoPath:
                    type: string
                    example: groups/1.png
                    description: the path of the group image
        "400":
          description: Invalid input image
          content:
            application/json:
              schema:
                description: A descriptive message of failure
                type: "object"
                properties:
                  message:
                    type: "string"
                    description: >
                      error:
                      * `First Error`  - No file uploaded.
                      * `Second Error` - Wrong file format. Only jpeg, png and webp are allowed.
                      * `Third Error`  - File too big. The image can't be bigger than 2Mb.
                    enum:
                      - no_file_uploaded
                      - wront_file_format
                      - file_too_big
                    default: "Wrong file format. Only jpeg, png and webp are allowed."
        "401":
          $ref: "#/components/responses/UnauthorizedError"
...
  /groups:
    post:
      tags:
        - Groups
      summary: Create group
      description: Create a new group with a name and members.
      operationId: createGroup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              description: Fields to create a group with
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 30
                  description: The name of the group
                members:
                  type: array
                  description: List of user ids to add to the group
                  items:
                    type: integer
            example:
              name: I ragazzi del calcetto
              members: [2, 6, 8]
      responses:
        "201":
          description: Group created successfully
          content:
            application/json:
              schema:
                description: A descriptive message of success and fields of the newly created group
                type: object
                properties:
                  groupId:
                    type: integer
                    description: The group id
                  name:
                    type: string
                    description: The group name
        "401":
          $ref: "#/components/responses/UnauthorizedError"